msvc doesn't define uint

+++ source/function.cpp
@@ -67,7 +67,7 @@ string function_arguments(clang::FunctionDecl const *record)
 {
 	string r;
 
-	for( uint i = 0; i < record->getNumParams(); ++i ) {
+	for( size_t i = 0; i < record->getNumParams(); ++i ) {
 		r += standard_name(record->getParamDecl(i)->getOriginalType().getCanonicalType().getAsString());
 		if( i + 1 != record->getNumParams() ) r += ", ";
 	}
@@ -81,11 +81,11 @@ string function_arguments(clang::FunctionDecl const *record)
 // Generate function argument list separate by comma
 // name_arguments - if arguments should be named: a1, a2, ...
 // n - number of arguments to generate. If n > num_of_function_parameters - generate only list with num_of_function_parameters
-pair<string, string> function_arguments_for_lambda(clang::FunctionDecl const *record, uint n)
+pair<string, string> function_arguments_for_lambda(clang::FunctionDecl const *record, size_t n)
 {
 	string r, a;
 
-	for( uint i = 0; i < record->getNumParams() and i < n; ++i ) {
+	for( size_t i = 0; i < record->getNumParams() and i < n; ++i ) {
 		QualType qt = record->getParamDecl(i)->getOriginalType().getCanonicalType();
 		r += standard_name(qt.getAsString()) + ' ';
 		if( !qt->isReferenceType() and !qt->isPointerType() ) r += !qt.isConstQualified() ? "const & " : "& ";
@@ -110,7 +110,7 @@ tuple<string, string, string> function_arguments_for_py_overload(clang::Function
 {
 	string r, a, p;
 
-	for( uint i = 0; i < record->getNumParams(); ++i ) {
+	for( size_t i = 0; i < record->getNumParams(); ++i ) {
 		QualType qt = record->getParamDecl(i)->getOriginalType().getCanonicalType();
 		r += standard_name(qt.getAsString()) + ' ' + "a" + std::to_string(i);
 		a += "a" + std::to_string(i);
@@ -156,7 +156,7 @@ string template_specialization(FunctionDecl const *F)
 		// if( FunctionDecl const *master = F->getTemplateInstantiationPattern() ) {
 		// 	outs() << "master for: " << F->getNameAsString() << "\n";
 		// 	if( TemplateArgumentList const *ta = F->getTemplateSpecializationArgs() ) {
-		// 		for(uint i=0; i < ta->size(); ++i) {
+		// 		for(size_t i=0; i < ta->size(); ++i) {
 		// 			string arg = template_argument_to_string( ta->get(i) );
 		// 			outs() << arg << " kind: " << ta->get(i).getKind()  << "\n";
 		// 		}
@@ -166,7 +166,7 @@ string template_specialization(FunctionDecl const *F)
 		if( TemplateArgumentList const *ta = F->getTemplateSpecializationArgs() ) {
 
 			templ += "<";
-			for( uint i = 0; i < ta->size(); ++i ) {
+			for( size_t i = 0; i < ta->size(); ++i ) {
 				// if( ta->get(i).getKind() != TemplateArgument::ArgKind::Null ) {
 				// outs() << "function: '" << F->getNameAsString() << "' template argument[" << i << "]=" << template_argument_to_string( ta->get(i) ) << " kind:" << ta->get(i).getKind() << "\n";
 
@@ -263,12 +263,12 @@ vector<QualType> get_type_dependencies(FunctionDecl const *F)
 	vector<QualType> r;
 
 	r.push_back(F->getReturnType()); //.getDesugaredType(F->getASTContext()) );
-	for( uint i = 0; i < F->getNumParams(); ++i ) r.push_back(F->getParamDecl(i)->getOriginalType() /*.getDesugaredType(F->getASTContext())*/);
+	for( size_t i = 0; i < F->getNumParams(); ++i ) r.push_back(F->getParamDecl(i)->getOriginalType() /*.getDesugaredType(F->getASTContext())*/);
 
 	// if( F->getTemplatedKind() == FunctionDecl::TK_MemberSpecialization  or   F->getTemplatedKind() == FunctionDecl::TK_FunctionTemplateSpecialization ) {
 	if( F->getTemplatedKind() != FunctionDecl::TK_NonTemplate ) {
 		if( TemplateArgumentList const *tal = F->getTemplateSpecializationArgs() ) {
-			for( uint i = 0; i < tal->size(); ++i ) {
+			for( size_t i = 0; i < tal->size(); ++i ) {
 				TemplateArgument const &ta(tal->get(i));
 				if( ta.getKind() == TemplateArgument::Type ) r.push_back(ta.getAsType());
 			}
@@ -320,7 +320,7 @@ bool is_skipping_requested(FunctionDecl const *F, Config const &config)
 
 
 // Generate binding for given function: .def("foo", (std::string (aaaa::A::*)(int) ) &aaaa::A::foo, "doc")
-string bind_function(FunctionDecl const *F, uint args_to_bind, bool request_bindings_f, Context &context, CXXRecordDecl const *parent, bool always_use_lambda)
+string bind_function(FunctionDecl const *F, size_t args_to_bind, bool request_bindings_f, Context &context, CXXRecordDecl const *parent, bool always_use_lambda)
 {
 	string function_name = python_function_name(F);
 
@@ -340,7 +340,7 @@ string bind_function(FunctionDecl const *F, uint args_to_bind, bool request_bind
 	}
 	else {
 		pair<string, string> args = function_arguments_for_lambda(F, args_to_bind);
-		// string args; for(uint i=0; i<args_to_bind; ++i) args += "a" + std::to_string(i) + ( i+1 == args_to_bind ? "" : ", " );
+		// string args; for(size_t i=0; i<args_to_bind; ++i) args += "a" + std::to_string(i) + ( i+1 == args_to_bind ? "" : ", " );
 
 		string return_type = standard_name(F->getReturnType().getCanonicalType().getAsString());
 
@@ -385,7 +385,7 @@ string bind_function(FunctionDecl const *F, uint args_to_bind, bool request_bind
 
 	if( request_bindings_f ) request_bindings(F->getReturnType().getCanonicalType(), context);
 
-	for( uint i = 0; i < F->getNumParams() and i < args_to_bind; ++i ) {
+	for( size_t i = 0; i < F->getNumParams() and i < args_to_bind; ++i ) {
 		r += ", pybind11::arg(\"{}\")"_format(string(F->getParamDecl(i)->getName()));
 
 		if( request_bindings_f ) request_bindings(F->getParamDecl(i)->getOriginalType(), context);
